% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.Acc_Process.Part1_ExtractRawAcc.R
\name{E4.Acc_Process.part1.ExtractRawAcc}
\alias{E4.Acc_Process.part1.ExtractRawAcc}
\title{Acc Processing Part 1: Extract raw acceleromter data}
\usage{
E4.Acc_Process.part1.ExtractRawAcc(participant_list, ziplocation,
  rdslocation.acc)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{ziplocation}{folder location where the participant-level subfolders are (make sure that it ends in /)}

\item{rdslocation.acc}{folder location where you want the RDS outputs to go (make sure that it ends in /)}
}
\description{
This allows you extract acceleromter data. It will output raw acceleromter data (x,y,z).
Inputs are: (1) List of participant numbers and (2) location where ZIP folders are stored. Outputs are: (1) one RDS file per participant with all data. A working example and vignette will be added later.
}
\examples{
E4.Acc_Process.part1.ExtractRawAcc(
  participant_list=c(1001),
  ziplocation=paste(system.file(package="E4tools"),
  "/extdata/E4_demo_data/",sep=""),
  rdslocation.acc=paste(tempdir(),"/extdata/output/raw_acc/",sep=""))
}
\keyword{acc}
