% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.Prep.FileHelper.R
\name{E4.Step0.FileHelper}
\alias{E4.Step0.FileHelper}
\title{Set global file locations to make other functions easier}
\usage{
E4.Step0.FileHelper(participant_list, ziplocation, dataroot)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{ziplocation}{folder location where the participant-level subfolders are (make sure that it ends in /)}

\item{dataroot}{folder where you want your data to be stored.}
}
\description{
This function will allow you to pre-define file locations that are used in multiple functions so you only have to type them once and so that your folder structure will be well-organized.
}
\examples{
\dontrun{E4.Acc_Process.Part1.ExtractRawAcc(participant_list=c(1001:1002),
ziplocation="~/documents/study/data/",
rdslocation.acc="~/documents/study/data/acc/")}
}
\keyword{acc}
