% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.extras.BinTemp.R
\name{E4.Temp.part2.bin_temp}
\alias{E4.Temp.part2.bin_temp}
\title{Temperature part 2: Make temperature bins}
\usage{
E4.Temp.part2.bin_temp(participant_list, rdslocation.temp,
  rdslocation.binnedtemp, BinLengthMin)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{rdslocation.temp}{folder location where raw temperature (from part 1) is saved.}

\item{rdslocation.binnedtemp}{folder location where you want the RDS outputs to go (make sure that it ends in /)}

\item{BinLengthMin}{folder location where you want the RDS outputs to go (make sure that it ends in /)}
}
\description{
Put temperature data in bins of X minutes length
}
\examples{
\dontrun{E4.extras.BinEDA(participant_list=c(1001:1004),rdslocation.binnedtemp="~/study/data/EDA/",
rdslocation.binnedtemp="~/study/data/Binned_EDA/",
BinLengthMin=2,
RejectFlag=TRUE)}
}
\keyword{acc}
