% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.part3.MatchPressesToEDA.R
\name{E4_EDA_Process.part3.MatchPressesToEDA}
\alias{E4_EDA_Process.part3.MatchPressesToEDA}
\title{Match EDA data to button pressess}
\usage{
E4_EDA_Process.part3.MatchPressesToEDA(participant_list,
  rdslocation.MatchedEDA, rdslocation.EDA, rdslocation.buttonpress,
  min.before, min.after)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{rdslocation.MatchedEDA}{folder location where you want the RDS outputs to go (make sure that it ends in /). The combined data file will go into this directory. Individual participants' data will go into a subdirectory in this folder called "individual_participants"}

\item{rdslocation.EDA}{folder where rds files for individual Ps' EDA data are stored (from part 1)}

\item{rdslocation.buttonpress}{location of folder where button press output is stored (from part 2)}

\item{min.before}{how many minutes before a button press do you want EDA data? Enter 0 if you do not want ANY data before (i.e., you're using only data post-press)}

\item{min.after}{how many minutes after a button press do you want EDA data? Enter 0 if you do not want ANY data after (i.e., you're using only data pre-press)}
}
\description{
This function allows you to extract the data that are within X minutes before and/or after a button press.
}
\examples{
\dontrun{E4_EDA_Process.part3.MatchPressesToEDA(participant_list=c(1001:1008,1011:1014,1017,1021),
rdslocation.buttonpress="/Users/documents/study/data/tags/",
rdslocation.MatchedEDA="/Users/documents/study/data/matched/",
rdslocation.EDA="/Users/documents/study/data/EDA/",
min.before=20,min.after=20)}




}
\keyword{EDA}
