% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4tools-package.R
\docType{package}
\name{E4tools-package}
\alias{E4tools}
\alias{E4tools-package}
\title{E4tools: Management and Processing Tools for Data Produced by the Empatica E4}
\description{
Process and manage the data from the Empatica E4. All functions operate on the EDA data stream, but other streams will be added soon. The Empatica E4 is a wearable physiological monitor made by Empatica (Empatica is not associated with any of this code). You can find more information about the E4 at Empatica's website <https://www.empatica.com/research/e4/>.
}
\author{
\strong{Maintainer}: Evan Kleiman \email{evan.kleiman@rutgers.edu}

}
\keyword{internal}
