% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.Diagnostics.EDAplot.R
\name{E4.Diagnostics.EDAplot}
\alias{E4.Diagnostics.EDAplot}
\title{Diagnostics: Plot EDA data and button presses}
\usage{
E4.Diagnostics.EDAplot(participant_list, rdslocation.binnedEDA,
  rdslocation.buttonpress, plotlocation.EDA, RejectFlagCount = 48)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{rdslocation.binnedEDA}{folder location where binned EDA is stored (from E4.extras.BinEDA function)}

\item{rdslocation.buttonpress}{location of folder where button press output is stored (from part 2)}

\item{plotlocation.EDA}{Folder where you want to store the PDF plots}

\item{RejectFlagCount}{What percent of samples in the bin must be bad for the entire bin to be marked bad? Default is 48, which is 10 percent of samples in a 2-minute bin.}
}
\description{
This will allow you to see all binned EDA data for a participant, along with which band they were wearing and when they pressed the event marker. One PDF file is made per participant.
}
\examples{
\dontrun{E4.Diagnostics.EDAplot(participant_list=c(1001:1004),
rdslocation.buttonpress="~/study/data/tags/",
rdslocation.binnedEDA="~/study/data/Binned_EDA/",
plotlocation.EDA="~/study/data/EDAplots/")}




}
\keyword{diagnostics}
\keyword{plots}
