% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4.part2.ExtractButtonPresses.R
\name{E4_EDA_Process.part2.ExtractButtonPresses}
\alias{E4_EDA_Process.part2.ExtractButtonPresses}
\title{EDA Processing Part 2: Extract button pressess}
\usage{
E4_EDA_Process.part2.ExtractButtonPresses(participant_list, ziplocation,
  rdslocation.buttonpress, part1summaries, cutoff.ends = 0,
  cutoff.overlap = 0)
}
\arguments{
\item{participant_list}{list of participant numbers NOTE: This should match the names of the folders (e.g., participant 1001's data should be in a folder called "1001")}

\item{ziplocation}{folder location where the participant-level subfolders are (make sure that it ends in /)}

\item{rdslocation.buttonpress}{folder location where you want the RDS output to go (make sure that it ends in /). The file will be named "button_presses.RDS"}

\item{part1summaries}{location of folder where summaries from part 1 were saved (make sure that it ends in /)}

\item{cutoff.ends}{how close (in minutes) to the ends of a file do you want to cut off button pressess (because they could be accidental e.g., when turing the band off). Default is 0, which will not remove button presess at all.}

\item{cutoff.overlap}{if you want to remove button pressess within X number of minutes, enter that value here. Default is 0, which will not remove button presess at all.}
}
\description{
This function allows you extract button pressess and remove pressess that are within a certain number of minutes before the end of a session or that are too close to another button press.
}
\examples{
\dontrun{E4_EDA_Process.part2.ExtractButtonPresses(participant_list=c(1001:1008,1011:1014,1017,1021),
ziplocation="/Users/documents/study/data/Raw_E4_Data/",
rdslocation.buttonpress="/Users/documents/study/data/tags/",
part1summaries="/Users/documents/study/data/EDA/summaries/",
cutoff.ends=2,
cutoff.overlap=20)}




}
\keyword{EDA}
