% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E4tools-package.R
\docType{package}
\name{E4tools-package}
\alias{E4tools}
\alias{E4tools-package}
\title{E4tools: Management and Processing Tools for Data Produced by the Emaptica E4}
\description{
This package allows you to process and manage the data from the Empatica E4. All functions operate on the EDA data stream, but other streams will be added soon.
}
\author{
\strong{Maintainer}: Evan Kleiman \email{evan.kleiman@rutgers.edu}

}
\keyword{internal}
